Refactored a lot of the database connection code and comments. This includes implementing a Settings table in the database to check for existence of the database, and house other dabase specific settings. Implemented a Database Connection Pool on the MariaDB/MySQL connection whereas I left the connection for SQLite a single connection as that is what was advised by the research I did. Thus I there should be a speedup while accessing a MariaDB/MySQL databasee over SQLite access.
