+++ build.gradle
@@ -24,7 +24,7 @@ ext {
     // Version 0.1 for including the sponge api and creating the initial first class.
     sponge_version = '0.1'
     // Version 0.4 from the WageTable Update for the dev build
-    common_version = '0.10'
+    common_version = '0.11'
     cp_build = 'dev'
     jar_root_prefix = 'craftyprof'
 }
