+++ common/build.gradle
@@ -27,8 +27,8 @@ repositories {
 dependencies {
     implementation group: 'com.google.code.gson', name: 'gson', version: '2.8.5'
     implementation group: 'com.google.guava', name: 'guava', version: '19.0'
-    implementation group: 'org.xerial', name: 'sqlite-jdbc', version: '3.23.1'
-    implementation group: 'org.mariadb.jdbc', name: 'mariadb-java-client', version: '2.2.5'
+    implementation group: 'org.xerial', name: 'sqlite-jdbc', version: '3.25.2'
+    implementation group: 'org.mariadb.jdbc', name: 'mariadb-java-client', version: '2.4.0'
 
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.3.2'
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.3.2'
