+++ spigot/src/main/java/com/creedfreak/spigot/CraftyProfessionsSpigot.java
@@ -5,8 +5,8 @@ import com.creedfreak.common.ICraftyProfessions;
 import com.creedfreak.common.commands.AbsCmdController;
 import com.creedfreak.common.container.PlayerManager;
 import com.creedfreak.common.container.WageTableHandler;
-import com.creedfreak.common.database.databaseConn.Database;
-import com.creedfreak.common.database.databaseConn.DatabaseFactory;
+import com.creedfreak.common.database.connection.Database;
+import com.creedfreak.common.database.connection.DatabaseFactory;
 import com.creedfreak.common.utility.Logger;
 import com.creedfreak.spigot.commands.CommandController;
 import com.creedfreak.spigot.config.ConfigController;
